Rename build_pipeline
