+++ src/scripts/build_dataset.py
@@ -7,7 +7,7 @@ Usage:
 import hydra
 from omegaconf import DictConfig
 
-from zebra_puzzles.pipeline import build_dataset
+from zebra_puzzles.build_pipeline import build_dataset
 
 
 @hydra.main(config_path="../../config", config_name="config", version_base=None)
